Moved VirtualRobotGui.
