+++ aima-gui/src/main/java/aima/gui/swing/demo/robotics/simple/VirtualRobotGui.java
@@ -1,7 +1,10 @@
-package aima.gui.swing.demo.robotics.components;
+package aima.gui.swing.demo.robotics.simple;
 
 import aima.core.robotics.impl.datatypes.Angle;
-import aima.gui.swing.demo.robotics.simple.VirtualRobot;
+import aima.gui.swing.demo.robotics.components.AnglePanel;
+import aima.gui.swing.demo.robotics.components.IRobotGui;
+import aima.gui.swing.demo.robotics.components.Settings;
+import aima.gui.swing.demo.robotics.components.AnglePanel.ChangeListener;
 import aima.gui.swing.framework.util.GuiBase;
 
 /**
